topic 2
point 9
