+++ matthieu/02-oop-inheritance/assets/js/Actor.js
@@ -1,0 +1,9 @@
+
+class Actor {
+  constructor(name, age) {
+    this.name = name;
+    this.age = age;
+  }
+}
+
+module.exports = Actor;
