+++ matthieu/02-oop-inheritance/assets/js/Logger.js
@@ -1,0 +1,8 @@
+class Logger {
+  constructor() {}
+  log(info) {
+    console.log("The " + info + " event has been emitted.");
+  }
+}
+
+module.exports = Logger;
