+++ matthieu/02-oop-inheritance/index.html
@@ -17,9 +17,9 @@
         <script src="assets/js/plugins.js"></script>
         <script src="assets/js/main.js"></script>
 
+         <script type="text/javascript" src="bundle.js"></script>
+		<!-- <script src="assets/js/index.js"></script>  -->
        
-		<script src="assets/js/index.js"></script>  
-
         <link href="assets/css/index.css" rel="stylesheet">
 		<link href="assets/css/normalize.css" rel="stylesheet">
