+++ matthieu/02-oop-inheritance/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "topic2",
+  "version": "1.0.0",
+  "description": "topic2",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "babelify": "^6.4.0",
+    "browserify": "^14.3.0",
+    "gulp": "^3.9.1",
+    "gulp-util": "^3.0.8",
+    "vinyl-source-stream": "^1.1.0"
+  }
+}
